make qrcode in terminal
checks for qrencode to be installed, install it if not. Then shows a generated image using eog and after closing the eog, it deletes the generated image. --name (-n) is also supported alongside --keep (-k) to keep the generated image with a chosen name. if no name is specified, the image is stored as temp